Agent.mcp/requirements.txt
mcp
python-dotenv
requests
fastapi
uvicorn[standard]
